Fix typos, rename classes package, rename Discord/MicrosoftInfo
