+++ bot/src/org/epilink/bot/http/DiscordUserInfo.kt
@@ -4,7 +4,7 @@ package org.epilink.bot.http
  * Represents Discord user information retrieved from the Discord API. This is manually filled in, so it does not
  * *actually* correspond do any real Discord endpoint.
  */
-data class DiscordInfo(
+data class DiscordUserInfo(
     val id: String,
     val username: String,
     val avatarUrl: String?
