+++ bot/src/org/epilink/bot/http/MicrosoftUserInfo.kt
@@ -4,4 +4,4 @@ package org.epilink.bot.http
  * Represents Microsoft user information retrieved from the Microsoft API. This is manually filled in, so it does not
  * *actually* correspond do any real Microsoft Graph endpoint.
  */
-data class MicrosoftInfo(val guid: String, val email: String)
+data class MicrosoftUserInfo(val guid: String, val email: String)
