+++ bot/src/org/epilink/bot/http/data/AdditionalRegistrationOptions.kt
@@ -1,4 +1,4 @@
-package org.epilink.bot.http.classes
+package org.epilink.bot.http.data
 
 // See the Api.md documentation file for more information
 data class AdditionalRegistrationOptions(
